move store libs back to separate dir
