+++ lib/parser.js
@@ -1,6 +1,6 @@
 import { parseFile as abletonParseFile } from 'ableton-parser';
 import { defaultInkFile } from './ink-file';
-import { getByPath } from './project-store';
+import { getByPath } from './store/project-store';
 import glob from 'glob';
 import { diffArrayTooSimple } from './utils/array';
