+++ lib/project.js
@@ -1,7 +1,7 @@
 import { basename } from 'path';
 import uuid from 'uuid/v4';
 import Git from 'nodegit';
-import * as projectStore from './project-store';
+import * as projectStore from './store/project-store';
 import {
   gitCheckAndInit,
   gitStatus,
