+++ lib/store/store.js
@@ -1,5 +1,5 @@
 import Conf from 'conf';
-import { config } from './config';
+import { config } from '../config';
 
 // just take the original ElectronStore, but remove the `electron.app` reference
 // <https://github.com/sindresorhus/electron-store/blob/f88a6463bcf7e720629e04b7f1336a133bf9bff3/index.js>
